requirements.txt
streamlit
python-dotenv
ragstack-ai
#beautifulsoup4
boto3